+++ process.py
@@ -49,6 +49,16 @@ class Task:
         return result
 
 
+def size(graph):
+    if not graph:
+        return 0
+    else :
+        s=0
+        for key in graph:
+            s=s+1+size(graph[key])
+        return s
+
+
 def process(folder_id, max_iter=None, n_jobs=None, verbose=False):
     if n_jobs is None:
         n_jobs = max(1, multiprocessing.cpu_count() - 1)
@@ -95,6 +105,14 @@ def process(folder_id, max_iter=None, n_jobs=None, verbose=False):
         sys.stdout.flush()
         if n_completed == n_tasks:
             break
+
+    s=0
+    for res in results:
+        s=s+size(res['graph'])
+    avg=s/len(results)
+    print("\nAverage Difference Graph Size: {0}".format(avg))
+
+
 
     with open(folder_id + ".json", "w") as output_file:
         json.dump({"items": results}, output_file)
